# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "description" : "defaultDeription",
    "version" : "1.0.0-oas3"
  },
  "servers" : [ {
    "url" : "https://0812dce4.ngrok.io"
  } ],
  "paths" : {
...
    "/8API/Screen/Masters/Core__Account/{Code}" : {
      "get" : {
        "tags" : [ "Masters" ],
        "description" : "the method returns the “JEC” master with field data from the “Account” Master",
        "parameters" : [ {
          "name" : "AccessToken",
          "in" : "header",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Code",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful opeartion",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_2"
                },
                "examples" : { }
              }
            }
          },
          "403" : {
            "description" : "Forbidden - Access to the resource by the provider is forbidden"
          },
          "404" : {
            "description" : "Not found - The requested resource is not found"
          },
          "502" : {
            "description" : "Provider server error - Something went wrong on the Provider or Endpoint’s server"
          }
        },
        "servers" : [ {
          "url" : "https://0812dce4.ngrok.io"
        } ]
      },
      "servers" : [ {
        "url" : "https://0812dce4.ngrok.io"
      } ]
    },
...
    "/8API/List/Masters" : {
      "get" : {
        "tags" : [ "Masters" ],
        "description" : "This method will return the list of masters along with the field details of the master tag.",
        "parameters" : [ {
          "name" : "AccessToken",
          "in" : "header",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          },
          "example" : "UB+vNzB2kI0stV0HMVYOwpam9kIL1t3vDdFTBxSwoqY="
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_4"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden - Access to the resource by the provider is forbidden"
          },
          "404" : {
            "description" : "Not found - The requested resource is not found"
          },
          "502" : {
            "description" : "Provider server error - Something went wrong on the Provider or Endpoint’s server"
          }
        },
        "servers" : [ {
          "url" : "https://0812dce4.ngrok.io"
        } ]
      },
      "servers" : [ {
        "url" : "https://0812dce4.ngrok.io"
      } ]
    },
    "/8API/Masters/Core__Warehouse" : {
      "post" : {
        "tags" : [ "Masters" ],
        "description" : "Saving the particular master data",
        "parameters" : [ {
          "name" : "AccessToken",
          "in" : "header",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          },
          "example" : "UB+vNzB2kI0stV0HMVYOwpam9kIL1t3vDdFTBxSwoqY="
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/body_1"
              },
              "examples" : {
                "0" : {
                  "value" : "{\n  \"data\": [\n    {\n      \"Alias\": \"SGW123\",\n      \"sName\": \"SGW123\",\n      \"sCode\": \"SGW123\"\n    }\n  ],\n  \"result\": 1,\n  \"message\": \"\",\n  \"url\": \"https://0812dce4.ngrok.io/8API/Screen/Masters/Core__Warehouse/SGW\"\n}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Succesfull operation",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_5"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden - Access to the resource by the provider is forbidden"
          },
          "404" : {
            "description" : "Not found - The requested resource is not found"
          },
          "502" : {
            "description" : "Provider server error - Something went wrong on the Provider or Endpoint’s server"
          }
        },
        "servers" : [ {
          "url" : "https://0812dce4.ngrok.io"
        } ]
      },
      "servers" : [ {
        "url" : "https://0812dce4.ngrok.io"
      } ]
    },
...
    "/8API/List/Masters/Core__Account" : {
      "get" : {
        "tags" : [ "Masters" ],
        "description" : "the method returns the masters list with field data from the “Account” tag.",
        "parameters" : [ {
          "name" : "fields",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "sCode%2CiStatus"
        }, {
          "name" : "AccessToken",
          "in" : "header",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          },
          "example" : "UB+vNzB2kI0stV0HMVYOwpam9kIL1t3vDdFTBxSwoqY="
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_8"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden - Access to the resource by the provider is forbidden"
          },
          "404" : {
            "description" : "Not found - The requested resource is not found"
          },
          "502" : {
            "description" : "Provider server error - Something went wrong on the Provider or Endpoint’s server"
          }
        },
        "servers" : [ {
          "url" : "https://0812dce4.ngrok.io"
        } ]
      },
      "servers" : [ {
        "url" : "https://0812dce4.ngrok.io"
      } ]
    },
...
    "/8API/Masters/Core__TestMaster/{MasterCode}" : {
      "delete" : {
        "tags" : [ "Masters" ],
        "description" : "For delete the Particular master",
        "parameters" : [ {
          "name" : "AccessToken",
          "in" : "header",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          },
          "example" : "lbMqgK25vHPk1cEI6rk7QqbJCX+7Hufv7+wu9TmcTV0="
        }, {
          "name" : "MasterCode",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Auto generated using Swagger Inspector",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_9"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden - Access to the resource by the provider is forbidden"
          },
          "404" : {
            "description" : "Not found - The requested resource is not found"
          },
          "502" : {
            "description" : "Provider server error - Something went wrong on the Provider or Endpoint’s server"
          }
        },
        "servers" : [ {
          "url" : "https://0812dce4.ngrok.io"
        } ]
      },
      "servers" : [ {
        "url" : "https://0812dce4.ngrok.io"
      } ]
    },
...
  },
  "components" : {
    "schemas" : {
...
      "inline_response_200_2_data" : {
        "type" : "object",
        "properties" : {
          "iAccountType" : {
            "type" : "integer"
          },
          "iDeliveryCity__Name" : {
            "type" : "string"
          },
          "iCreatedDate" : {
            "type" : "string"
          },
          "Alias" : {
            "type" : "string"
          },
          "iStatus" : {
            "type" : "integer"
          },
          "sCode" : {
            "type" : "string"
          },
          "BinCapacity" : {
            "type" : "integer"
          },
          "iPDCDiscountedAC__Name" : {
            "type" : "string"
          },
          "sName" : {
            "type" : "string"
          }
        },
        "example" : {
          "iDeliveryCity__Name" : "iDeliveryCity__Name",
          "sCode" : "sCode",
          "iCreatedDate" : "iCreatedDate",
          "Alias" : "Alias",
          "sName" : "sName",
          "iAccountType" : 6,
          "iStatus" : 1,
          "iPDCDiscountedAC__Name" : "iPDCDiscountedAC__Name",
          "BinCapacity" : 5
        }
      },
      "inline_response_200_2" : {
        "type" : "object",
        "properties" : {
          "result" : {
            "type" : "integer"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/inline_response_200_2_data"
            }
          },
          "message" : {
            "type" : "string"
          }
        },
        "example" : {
          "result" : 0,
          "data" : [ {
            "iDeliveryCity__Name" : "iDeliveryCity__Name",
            "sCode" : "sCode",
            "iCreatedDate" : "iCreatedDate",
            "Alias" : "Alias",
            "sName" : "sName",
            "iAccountType" : 6,
            "iStatus" : 1,
            "iPDCDiscountedAC__Name" : "iPDCDiscountedAC__Name",
            "BinCapacity" : 5
          }, {
            "iDeliveryCity__Name" : "iDeliveryCity__Name",
            "sCode" : "sCode",
            "iCreatedDate" : "iCreatedDate",
            "Alias" : "Alias",
            "sName" : "sName",
            "iAccountType" : 6,
            "iStatus" : 1,
            "iPDCDiscountedAC__Name" : "iPDCDiscountedAC__Name",
            "BinCapacity" : 5
          } ],
          "message" : "message"
        }
      },
...
      "inline_response_200_4_data" : {
        "type" : "object",
        "properties" : {
          "APIMasterName" : {
            "type" : "boolean"
          },
          "MasterTypeId" : {
            "type" : "integer"
          },
          "Caption" : {
            "type" : "string"
          }
        },
        "example" : {
          "APIMasterName" : true,
          "Caption" : "Caption",
          "MasterTypeId" : 6
        }
      },
      "inline_response_200_4" : {
        "type" : "object",
        "properties" : {
          "result" : {
            "type" : "integer"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/inline_response_200_4_data"
            }
          },
          "message" : {
            "type" : "string"
          }
        },
        "example" : {
          "result" : 0,
          "data" : [ {
            "APIMasterName" : true,
            "Caption" : "Caption",
            "MasterTypeId" : 6
          }, {
            "APIMasterName" : true,
            "Caption" : "Caption",
            "MasterTypeId" : 6
          } ],
          "message" : "message"
        }
      },
      "8APIMastersCore__Warehouse_data" : {
        "type" : "object",
        "properties" : {
          "sCode" : {
            "type" : "string"
          },
          "Alias" : {
            "type" : "string"
          },
          "sName" : {
            "type" : "string"
          }
        }
      },
      "body_1" : {
        "type" : "object",
        "properties" : {
          "result" : {
            "type" : "integer"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/8APIMastersCore__Warehouse_data"
            }
          },
          "message" : {
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          }
        }
      },
      "inline_response_200_5_data" : {
        "type" : "object",
        "properties" : {
          "MasterId" : {
            "type" : "integer"
          },
          "VoucherNo" : {
            "type" : "string"
          },
          "Exception" : {
            "type" : "string"
          }
        },
        "example" : {
          "MasterId" : 6,
          "VoucherNo" : "VoucherNo",
          "Exception" : "Exception"
        }
      },
      "inline_response_200_5" : {
        "type" : "object",
        "properties" : {
          "result" : {
            "type" : "integer"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/inline_response_200_5_data"
            }
          },
          "message" : {
            "type" : "string"
          }
        },
        "example" : {
          "result" : 0,
          "data" : [ {
            "MasterId" : 6,
            "VoucherNo" : "VoucherNo",
            "Exception" : "Exception"
          }, {
            "MasterId" : 6,
            "VoucherNo" : "VoucherNo",
            "Exception" : "Exception"
          } ],
          "message" : "message"
        }
      },
...
      "inline_response_200_8_data" : {
        "type" : "object",
        "properties" : {
          "sCode" : {
            "type" : "string"
          },
          "iStatus" : {
            "type" : "integer"
          }
        },
        "example" : {
          "sCode" : "sCode",
          "iStatus" : 6
        }
      },
      "inline_response_200_8" : {
        "type" : "object",
        "properties" : {
          "result" : {
            "type" : "integer"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/inline_response_200_8_data"
            }
          },
          "message" : {
            "type" : "string"
          }
        },
        "example" : {
          "result" : 0,
          "data" : [ {
            "sCode" : "sCode",
            "iStatus" : 6
          }, {
            "sCode" : "sCode",
            "iStatus" : 6
          } ],
          "message" : "message"
        }
      },
      "inline_response_200_9" : {
        "type" : "object",
        "properties" : {
          "result" : {
            "type" : "integer"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "type" : "object"
            }
          },
          "message" : {
            "type" : "string"
          }
        },
        "example" : {
          "result" : 0,
          "data" : [ { }, { } ],
          "message" : "message"
        }
      },
...
    }
  }
}